fix: pass turtle_tests/{window_y_trigger, window_y_trigger_wx_offscreen} tests
Fixes to pass turtle_tests suite.
https://github.com/c-sp/game-boy-test-roms/tree/8e1f6d7f3a1d8683f11fdf23008d1b1b26e51b52/src/turtle-tests-expected

window_y_trigger, window_y_trigger_wx_offscreen fixes:
* Change the logic to determine how the window should be drawn in
the frame. Previously I was just comparing window.y and the line. But
according to pandocs we need to check if at some point in the frame the
value of WY was equal to LY.
Also, window Y position is now a state, rather than a value that can be
calculated when rendering a scanline.

* Remove unnecessary build dependencies from flake.nix.